Trait ic_types::artifact::ArtifactKind

pub trait ArtifactKind: Sized {
    type Id;
    type Message: Unpin;
    type SerializeAs;
    type Attribute;
    type Filter: Default;

    const TAG: ArtifactTag;

    fn to_advert(msg: &Self::Message) -> Advert<Self>;

    fn check_advert(
        msg: &Self::Message, 
        advert: &Advert<Self>
    ) -> Result<(), Advert<Self>>
    where
        Advert<Self>: Eq,
    { ... }
}

Related artifact sub-types (Message/Id/Attribute/Filter) are parameterized by a type variable, which is of ArtifactKind trait. It is mostly a convenience to pass around a collection of types instead of all of them individually.

Associated Types

type Id

type Message: Unpin

type SerializeAs

type Attribute

type Filter: Default

Loading content...
Associated Constants

const TAG: ArtifactTag

Loading content...
Required methods

fn to_advert(msg: &Self::Message) -> Advert<Self>

Returns the advert of the given message.

Loading content...
Provided methods

fn check_advert(
msg: &Self::Message,
advert: &Advert<Self>
) -> Result<(), Advert<Self>> where
Advert<Self>: Eq,

Checks if the given advert matches what is computed from the message. Returns the advert derived from artifact on mismatch.

Loading content...