Module ic_types::batch

Contains Batch, Payload, and specific Payload types that are passed between Consensus and Message Routing.

Structs

Batch
The Batch provided to Message Routing for deterministic processing.

BatchPayload
The payload of a batch.

IngressPayload
Payload that contains Ingress messages

ValidationContext
The context built by Consensus for deterministic processing. Captures all fields that have semantic meaning within the Chain Consensus protocol.

XNetPayload
Payload that contains XNet messages.

Enums

IngressPayloadError
Possible errors when accessing messages in an IngressPayload.

InvalidIngressPayload
Possible errors when converting from an IngressPayload to a Vec<SignedIngress>.

Type Definitions

IngressAndXNetMessages
Return ingress messages, xnet messages, and consensus responses.