Module ic_types::consensus::catchup

Defines types that allow outdated replicas to catch up to the latest state.

Structs

CUPWithOriginalProtobuf
A catch up package paired with the original protobuf. Note that the protobuf contained in this struct is only partially deserialized and has the ORIGINAL bytes CatchUpContent bytes that were signed in yet to be deserialized form.

CatchUpContentProtobufBytes
CatchUpContentProtobufBytes holds bytes that represent a protobuf serialized catch-up package

CatchUpContentT
A generic struct shared between CatchUpContent and CatchUpContentShare.

CatchUpPackageParam
The parameters used to request CatchUpPackage (by nodemanager).

Type Definitions

CatchUpContent
CatchUpContent contains all necessary data to bootstrap a subnet's participant.

CatchUpContentHash
CatchUpContentHash is the type of a hashed CatchUpContent

CatchUpPackage
CatchUpPackage is signed by a threshold public key. Its CatchUpContent is only trusted if the threshold public key is trusted.

CatchUpPackageShare
CatchUpPackageShare is signed by individual members in a threshold committee.

CatchUpShareContent
Content of CatchUpPackageShare uses the block hash to keep its size small.