Type Definition ic_types::consensus::catchup::CatchUpContent

type CatchUpContent = CatchUpContentT<HashedBlock>;

CatchUpContent contains all necessary data to bootstrap a subnet's participant.

Implementations

impl CatchUpContent

pub fn new(
block: HashedBlock,
random_beacon: HashedRandomBeacon,
state_hash: CryptoHashOfState
) -> Self

Create a new CatchUpContent

pub fn registry_version(&self) -> RegistryVersion

Return the registry version as recorded in the DKG summary of the block contained in the CatchUpContent.

pub fn from_share_content(share: CatchUpShareContent, block: Block) -> Self

Create a CatchupContent from a