Module ic_types::consensus::certification

Defines types used for certification.

Structs

Certification
A Certification is a CertificationContent that is cryptographically signed by a subnet using a threshold signature

CertificationContent
CertificationContent holds the data signed by certification

CertificationShare
A certification share is the signature of a single replica on a CertificationContent

Enums

CertificationMessage
CertificationMessage captures the different types of messages sent around for the purpose of state certification.

CertificationMessageHash
CertificationMessageHash contains the hash of a CertificationMessage.