Module ic_types::consensus::dkg

Defines types used for distributed key generation.

Structs

DealingContent
Holds the content of a DKG dealing

Dealings
Dealings contains dealing messages and the height at which this DKG interval started

Summary
The DKG summary will be present as the DKG payload at every block, corresponding to the start of a new DKG interval.

Enums

Payload
The DKG payload is either the DKG Summary, if this payload belongs to a start block of a new DKG interval, or a tuple containing the start height and the set of valid dealings corresponding to the current interval.

Type Definitions

DealingMessages
DealingMessages is a vector of DKG messages

Message
Contains a Node's contribution to a DKG dealing.