Module ic_types::consensus::hashed

An object that is hashable can be bundled together with its hash, which becomes a Hashed<H, V> object parameterized by the actual hash type H and value type V. There are a number of benefits of doing this:

Equality check can be done just on the hash, which is often cheaper than traversing the object recursively.

Computing hashes over Hashed<H, V> can also take a shortcut by just hashing the hash of type H, instead of the value of type V. Note that this does not lead to the same hash result as directly hashing the value.

Using Hash also gives a strawman implementation of Ord if the order does not have to be consistent with the actual order of the object's value.

The serialization of Hashed<H, V> will serialize both the hash and the value.

Structs

Hashed
Bundle of both a value and its hash. Once created it remains immutable, which is why both fields are only accessible through member functions, not as record fields.