Module ic_types::consensus::thunk

A Thunk<T> represents delayed initialization of a value of type T. That is, an initialization function of type FnOnce() -> T can be used to create a Thunk<T> object, and only called once when the thunk is evaluated. Evaluating the same thunk more than once will return the same value as its first evaluation, but will not result in calling the initialization function more than once.

An alternative way of initializing a thunk is to use From<T> trait.

To evaluate a thunk, we can either use the into_inner function or AsRef<T> trait.

The serialization of a Thunk<T> object will force an evaluation.

Implementation wise, Thunk is just a thin wrapper around once_cell::Lazy type, providing From, Serialize, and Deserialize trait implementations.

Structs

Thunk
A lazily initialized value of type T that is only initialized upon first evaluation.