Struct ic_types::consensus::thunk::Thunk

pub struct Thunk<T> { /* fields omitted */ }

A lazily initialized value of type T that is only initialized upon first evaluation.

Implementations

impl<T> Thunk<T>

pub fn new(init: Box<dyn FnOnce() -> T + Send>) -> Self

Return a Thunk<T> object with an initialization function init that will not be called until the thunk is evaluated.

pub fn into_inner(self) -> T

Convert a Thunk<T> object into its inner value of type T. It will force an evaluation if necessary.