Trait ic_types::consensus::ContentEq

pub trait ContentEq {
    fn content_eq(&self, other: &Self) -> bool;
}

Useful to compare equality by content, for example Signed<C,S> can be compared by equality on C.

Required methods

fn content_eq(&self, other: &Self) -> bool

content_eq compares two values and returns true if their content is equal. We implement this for signed artifacts, and say that they are content_eq if the value they sign is equal, even if the signature is different.

Loading content...