Module ic_types::crypto::dkg

Defines interactive distributed key generation (DKG) types.

Structs

Config
Configures interactive DKG.

Dealers
A set of dealers for interactive DKG.

Dealing
A dealing for interactive DKG.

DkgConfig
A validated configuration for interactive DKG. This configuration can only exist if all configuration invariants are satisfied. See DkgConfig::new for a description of the invariants.

DkgConfigData
The non-validated config parameter object to be passed to the DkgConfig constructor.

EncryptionPublicKey
An encryption public key for interactive DKG.

EncryptionPublicKeyPop
The encryption public key proof of posession.

EncryptionPublicKeyWithPop
An encryption public key together with its proof of posession.

Receivers
A set of receivers for interactive DKG.

Response
The dealer's response in the interactive DKG protocol. See DkgAlgorithm::create_response.

Transcript
An interactive DKG transcript.

TranscriptBytes
An interactive DKG transcript as bytes.