Enum ic_types::crypto::CryptoError

pub enum CryptoError {
    InvalidArgument {
        message: String,
    },
    PublicKeyNotFound {
        node_id: NodeId,
        key_purpose: KeyPurpose,
        registry_version: RegistryVersion,
    },
    TlsCertNotFound {
        node_id: NodeId,
        registry_version: RegistryVersion,
    },
    SecretKeyNotFound {
        algorithm: AlgorithmId,
        key_id: KeyId,
    },
    TlsSecretKeyNotFound {
        certificate_der: Vec<u8>,
    },
    MalformedSecretKey {
        algorithm: AlgorithmId,
        internal_error: String,
    },
    MalformedPublicKey {
        algorithm: AlgorithmId,
        key_bytes: Option<Vec<u8>>,
        internal_error: String,
    },
    MalformedSignature {
        algorithm: AlgorithmId,
        sig_bytes: Vec<u8>,
        internal_error: String,
    },
    MalformedPop {
        algorithm: AlgorithmId,
        pop_bytes: Vec<u8>,
        internal_error: String,
    },
    SignatureVerification {
        algorithm: AlgorithmId,
        public_key_bytes: Vec<u8>,
        sig_bytes: Vec<u8>,
        internal_error: String,
    },
    PopVerification {
        algorithm: AlgorithmId,
        public_key_bytes: Vec<u8>,
        pop_bytes: Vec<u8>,
        internal_error: String,
    },
    InconsistentAlgorithms {
        algorithms: BTreeSet<AlgorithmId>,
        key_purpose: KeyPurpose,
        registry_version: RegistryVersion,
    },
    AlgorithmNotSupported {
        algorithm: AlgorithmId,
        reason: String,
    },
    RegistryClient(RegistryClientError),
    ThresholdSigDataNotFound {
        dkg_id: DkgId,
    },
    DkgTranscriptNotFound {
        subnet_id: SubnetId,
        registry_version: RegistryVersion,
    },
    RootSubnetPublicKeyNotFound {
        registry_version: RegistryVersion,
    },
}

An error returned by the crypto component.

Variants

InvalidArgument

The arguments are semantically incorrect. This error is not retriable. This is equivalent to HTTP 422: The request was well-formed but was unable to be followed due to semantic errors.

Fields of InvalidArgument

message: String

PublicKeyNotFound

Public key for given (entity, purpose) pair not found at given registry version.

Fields of PublicKeyNotFound

node_id: NodeIdkey_purpose: KeyPurposeregistry_version: RegistryVersion

TlsCertNotFound

TLS cert for given node_id not found at given registry version.

Fields of TlsCertNotFound

node_id: NodeIdregistry_version: RegistryVersion

SecretKeyNotFound

Secret key not found in SecretKeyStore.

Fields of SecretKeyNotFound

algorithm: AlgorithmIdkey_id: KeyId

TlsSecretKeyNotFound

TLS secret key not found in SecretKeyStore.

Fields of TlsSecretKeyNotFound

certificate_der: Vec<u8>

MalformedSecretKey

Secret key could not be parsed or is otherwise invalid.

Fields of MalformedSecretKey

algorithm: AlgorithmIdinternal_error: String

MalformedPublicKey

Public key could not be parsed or is otherwise invalid.

Fields of MalformedPublicKey

algorithm: AlgorithmIdkey_bytes: Option<Vec<u8>>internal_error: String

MalformedSignature

Signature could not be parsed or is otherwise invalid.

Fields of MalformedSignature

algorithm: AlgorithmIdsig_bytes: Vec<u8>internal_error: String

MalformedPop

Pop could not be parsed or is otherwise invalid.

Fields of MalformedPop

algorithm: AlgorithmIdpop_bytes: Vec<u8>internal_error: String

SignatureVerification

Signature could not be verified.

Fields of SignatureVerification

algorithm: AlgorithmIdpublic_key_bytes: Vec<u8>sig_bytes: Vec<u8>internal_error: String

PopVerification

Pop could not be verified.

Fields of PopVerification

algorithm: AlgorithmIdpublic_key_bytes: Vec<u8>pop_bytes: Vec<u8>internal_error: String

InconsistentAlgorithms

Multi-signature: inconsistent (multiple) algorithms.

Fields of InconsistentAlgorithms

algorithms: BTreeSet<AlgorithmId>key_purpose: KeyPurposeregistry_version: RegistryVersion

AlgorithmNotSupported

Algorithm not supported.

Fields of AlgorithmNotSupported

algorithm: AlgorithmIdreason: String

RegistryClient(RegistryClientError)

Error querying the registry.

ThresholdSigDataNotFound

Threshold signature data store did not contain the expected data (public coefficients and node indices)

Fields of ThresholdSigDataNotFound

dkg_id: DkgId

DkgTranscriptNotFound

DKG transcript for given subnet ID not found at given registry version.

Fields of DkgTranscriptNotFound

subnet_id: SubnetIdregistry_version: RegistryVersion

RootSubnetPublicKeyNotFound

Root subnet public key not found at given registry version.

Fields of RootSubnetPublicKeyNotFound

registry_version: RegistryVersion

Implementations

impl CryptoError

pub fn is_public_key_not_found(&self) -> bool

pub fn is_secret_key_not_found(&self) -> bool

pub fn is_malformed_secret_key(&self) -> bool

pub fn is_malformed_public_key(&self) -> bool

pub fn is_malformed_signature(&self) -> bool

pub fn is_signature_verification_error(&self) -> bool

pub fn is_pop_verification_error(&self) -> bool

pub fn is_inconsistent_algorithms(&self) -> bool

pub fn is_algorithm_not_supported(&self) -> bool

pub fn is_registry_client_error(&self) -> bool

pub fn is_threshold_sig_data_not_found(&self) -> bool

pub fn is_dkg_transcript_not_found(&self) -> bool