Module ic_types::crypto::threshold_sig::ni_dkg

Types for non-interactive distributed key generation (NI-DKG).

Re-exports

pub use crate::crypto::threshold_sig::ni_dkg::config::receivers::NiDkgReceivers;

pub use id::NiDkgId;

Modules

config
Types related to the non-interactive DKG configuration.

errors
Errors for non-interactive DKG.

id
Types related to the non-interactive DKG ID.

transcripts_to_retain
Types related to transcripts that should be retained during NI-DKG key deletion. See NiDkgAlgorithm::retain_only_active_keys.

Structs

NiDkgDealing
A dealer's contribution (called dealing) to distributed key generation.

NiDkgTagIter

NiDkgTargetId
An ID for a remote NiDkgTargetSubnet.

NiDkgTranscript
Summarizes a distributed key generation.

Enums

DkgId
An ID identifying a DKG epoch.

NiDkgTag
Allows to distinguish protocol executions in high and low threshold settings.

NiDkgTargetSubnet
The subnet for which the DKG generates keys.