Module ic_types::crypto::threshold_sig::ni_dkg::config

Types related to the non-interactive DKG configuration.

Modules

dealers
Types related to dealers for non-interactive DKG.

errors
Errors that may occur in the context of validating non-interactive DKG configurations.

receivers
Types related to receivers for non-interactive DKG.

Structs

NiDkgConfig
A validated configuration for non-interactive DKG. This configuration can only exist if all configuration invariants are satisfied. See NiDkgConfig::new for a description of the invariants.

NiDkgConfigData
The non-validated config parameter object to be passed to the NiDkgConfig constructor.

NiDkgThreshold
The minimum number of nodes required to generate a valid threshold signature.