Module ic_types::crypto::threshold_sig::ni_dkg::errors

Errors for non-interactive DKG.

Modules

create_dealing_error
Errors related to dealing creation.

create_transcript_error
Errors related to transcript creation.

key_removal_error
Errors related to key removal.

load_transcript_error
Errors related to loading the transcript.

transcripts_to_retain_validation_error
Errors related to transcripts that should be retained.

verify_dealing_error
Errors related to verifying a dealing.

Structs

FsEncryptionPublicKeyNotInRegistryError
Occurs if the forward-secure encryption public key cannot be found in the registry.

MalformedFsEncryptionPublicKeyError
Occurs if the forward-secure encryption public key is malformed.

NotADealerError
Occurs if a node ID that should be a dealer is not a dealer.