Module ic_types::filetree_sync

File tree sync artifact.

Structs

FileTreeSyncArtifact
Artifact to be be delivered to the artifact pool when file tree sync is complete

FileTreeSyncChunksTracker
File tree sync tracker.

Enums

UnderConstructionState
Represents the state under construction for a file tree sync artifact.

Type Definitions

FileTreeSyncId
Unique identifier to describe a FSTreeSyncObject