Function ic_types::freeze_threshold_cycles

pub fn freeze_threshold_cycles(
    freeze_threshold: NumSeconds, 
    gib_storage_per_second_fee: Cycles, 
    expected_canister_size: NumBytes
) -> Cycles

Figure out how many cycles a canister should have so that it can support the given amount of storage for the given amount of time, given the storage fee.