Module ic_types::methods

This module contains a collection of types and structs that define the various types of methods in the IC.

Structs

Callback
References to functions executed when a response is received.

WasmClosure
A Wasm closure pointing to the Wasm function table.

Enums

FuncRef
A reference to a callable function/method in a Wasm module, which can be:

SystemMethod
The various system methods available to canisters.

WasmMethod
Represents the types of methods that a Wasm module can export.