Module ic_types::p2p

Defines types used by the P2P component.

Structs

GossipAdvert
This is sent to peers to indicate that a node has a certain artifact in its artifact pool. The adverts of different artifact types may differ in their attributes. Upon the reception of an advert, a node can decide if and when to request the corresponding artifact from the sender.

Constants

MAX_ARTIFACT_STREAMS_PER_PEER
Maximum number of artifact chunks that can be downloaded simultaneously from one peer

MAX_CHUNK_SIZE
Maximum size in bytes of an artifact chunk. Used to compute the chunk timeout interval.

MAX_CHUNK_WAIT_MS
Timeout interval (in milliseconds) within which a chunk request must succeed

MAX_DUPLICITY
Maximum number of peers that one artifact chunk can be downloaded from in parallel (MAX_DUPLICITY=1 means no parallel downloads)

PFN_EVALUATION_PERIOD_MS
Period for priority function evaluation in milliseconds

RECEIVE_CHECK_PEER_SET_SIZE
Size of each receive check hash set for each peer

REGISTRY_POLL_PERIOD_MS
Period for polling the registry for changes in milliseconds

RETRANSMISSION_REQUEST_MS
Period for sending a retransmission request in milliseconds

Functions

build_default_gossip_config
Helper function to build a gossip config using default values.