Module ic_types::registry::connection_endpoint

Defines ConnectionEndpoint as a URL.

Structs

ConnectionEndpoint
An endpoint is completely defined by a URL.

Enums

ConnectionEndpointTryFromError
Errors that can occur when converting from a URL to a ConnectionEndpoint.

ConnectionEndpointTryFromProtoError
Errors that can occur when converting from the protobuf encoding to the ConnectionEndpoint type.

ConnectionEndpointTryFromStringError
Errors that can occur when converting from a string to a ConnectionEndpoint.

Protocol