Module ic_types::state_sync

State sync types.

Note [Manifest Hash]

This note describes how hashes included into the manifest are computed.

All integers are hashed using their Big-Endian representation. E.g., u64 occupies 8 bytes and 1 is encoded as 0000_0000_0000_0001, and u32 occupies 4 bytes and 1 is encoded as 0000_0001.

All the variable-size values (paths, arrays) are prefixed with their length to get a unique encoding. If it wasn't the case, it would be possible to interpret 2 file table entries

[ "path1" size1 hash1 ]
[ "path2" size2 hash2 ]

as a single file table entry:

[ "path1 size1 hash1 path2" size2 hash2 ]

and get the same manifest hash.

Below are the rules for computing hashes of table entries and the root hash of a manifest:

The hash in the chunk table is simply the hash of the raw chunk content.

  chunk_hash := hash(dsep("ic-state-chunk") · file[offset:offset + size_bytes])

The hash in the file table is the hash of the slice of the chunk table corresponding to this file:

  file_hash   := hash(dsep("ic-state-file") · len(slice) as u32 · chunk_entry*)
  chunk_entry := file_index as u32
                 · size_bytes as u32
                 · offset as u64
                 · chunk_hash

The hash of the whole manifest is computed by hashing the file table:

  manifest_hash := hash(dsep("ic-state-manifest")
                   · version as u32
                   · len(file_table) as u32
                   · file_entry*
                   · len(chunk_table) as u32
                   · chunk_entry*
                   )
  file_entry    := len(relative_path) as u32
                   · relative_path
                   · size_bytes as u64
                   · file_hash

where

dsep(seq) = byte(len(seq)) · seq
``

Modules

proto
Conversions from Rust to proto structs and back for StateSync.

Structs

ChunkInfo
An entry of the chunk table.

FileInfo
An entry of the file table.

Manifest
Manifest is a short description of the checkpoint contents.

Constants

MANIFEST_CHUNK
Id of the manifest chunk in StateSync artifact.

Functions

decode_manifest
Deserializes the manifest from a byte array.

encode_manifest
Serializes the manifest into a byte array.