Struct ic_types::PrincipalId

pub struct PrincipalId { /* fields omitted */ }

The type representing principals as described in the interface spec.

A principal is just a blob that is displayed in a particular way. (see https://sdk.dfinity.org/docs/interface-spec/index.html#textual-ids)

Principals have variable length, bounded by 29 bytes. Since we want PrincipalId to implement the Copy trait, we encode them as a fixed-size array and a length.

Implementations

impl PrincipalId

pub const MAX_LENGTH_IN_BYTES: usize

pub fn as_slice(&self) -> &[u8]

pub fn to_vec(&self) -> Vec<u8>

pub fn into_vec(self) -> Vec<u8>

impl PrincipalId

Some principal ids have special classes (system-generated, self-authenticating, derived), see https://sdk.dfinity.org/docs/interface-spec/index.html#id-classes

The following functions allow creating and testing for the special forms.

pub fn new_user_test_id(n: u64) -> PrincipalId

pub fn new_node_test_id(n: u64) -> PrincipalId

pub fn new_subnet_test_id(n: u64) -> PrincipalId

pub const fn new(len: usize, data: [u8; 29]) -> PrincipalId

pub fn new_self_authenticating(pubkey: &[u8]) -> PrincipalId

pub fn new_derived(registerer: &PrincipalId, seed: &[u8]) -> PrincipalId

pub fn new_anonymous() -> PrincipalId

pub fn authenticates_for_pubkey(&self, pubkey: &[u8]) -> bool

pub fn is_self_authenticating(&self) -> bool

pub fn is_derived(&self, registerer: &PrincipalId, seed: &[u8]) -> bool

pub fn is_anonymous(&self) -> bool