Trait ic_types::CountBytes

pub trait CountBytes {
    fn count_bytes(&self) -> usize;
}

Allow an object to report its own byte size. It is only meant to be an estimate, and not an exact measure of its heap usage or length of serialized bytes.

Required methods

fn count_bytes(&self) -> usize

Loading content...