Module ic_types::transport

Transport layer public types.

Structs

Bytes
Represents a blob.

FlowId
FlowId is the unique key for the flows being managed

FlowTagType

TransportConfig
The transport format specified in the ic.json

TransportFlowConfig
Per-flow config

TransportFlowInfo
Information that can be used to identify a transport flow.

TransportPayload
The payload for the transport layer.

Enums

TransportClientType
The type of a transport client.

TransportError
Errors that are returned by the transport layer.

TransportErrorCode
Error codes returned by transport manager functions.

TransportNotification
A transport notification.

TransportStateChange
State changes that can happen in the transport layer.

Type Definitions

FlowTag
A tag attached to a flow.