Enum ic_types::transport::TransportErrorCode

pub enum TransportErrorCode {
    TransportClientAlreadyRegistered,
    TransportClientNotFound,
    PeerAlreadyRegistered,
    PeerNotFound,
    FlowAlreadyEnabled,
    FlowNotFound,
    FlowConnectionUp,
    TransportClientConfigNotFound,
    TransportBusy(TransportPayload),
    ConnectionWriteFailed(String),
    ConnectionReadFailed(String),
    SerializationFailed,
    DeserializationFailed,
    RegistryMissingConfig,
    RegistryMultiIP,
    RegistryInvalidPortNumber,
    MessageQueueRoutingFailed,
    TransportQueueFull,
    TransportQueueStopped,
    ConnectionEventHandlerNotFound,
    MessageSerializationFailed,
    MessageConnectionRoutingFailed,
    ConnectionNotFound,
    SocketWriteFailed,
    ServerSocketCreateFailed,
    ServerSocketAddrReuseFailed,
    ServerSocketPortReuseFailed,
    ServerSocketBindFailed,
    ServerSocketListenFailed,
    ServerSocketConversionFailed,
    ClientSocketCreateFailed,
    ClientSocketBindFailed,
    SocketNoDelayFailed,
    ServerDown,
    ConnectOsError,
    RegistryDuplicateNodeId,
    RegistryDuplicateNodeIP,
    RegistryDuplicateEndpoint,
    RegistryInvalidNodeIP,
    NodeIpResolutionFailed,
    NodeServerEndpointResolutionFailed,
    WrapperCertParsingFailed,
    NodeIdMissing,
    NodeIdParsingFailed,
    NodeIdMalformed,
    DomainNameMissing,
    DomainNameTooMany,
    DomainNameMalformed,
    PublicKeyParsingFailed,
    RegistryTlsConfigNotUniform,
    InvalidNodeIdInCertificate,
    PeerTlsInfoNotFound,
    PeerTlsInfoMismatch,
    ConfigPrivateKeyParsingFailed,
    ConfigPrivateKeyFileReadFailed,
    ConfigCertParsingFailed,
    ConfigCertFileReadFailed,
    SetNodeKeyFailed,
    SetNodeCertFailed,
    AddPeerCertFailed,
    AcceptorInitFailed,
    ConnectorInitFailed,
    ConnectorConfigFailed,
    SenderErrorIndicated,
    InvalidSockAddr,
    NodeRecordDuplicateFlowTag,
    NodeRecordMissingConnectionEndpoint,
    TimeoutExpired,
}

Error codes returned by transport manager functions.

Variants

TransportClientAlreadyRegistered

Found an active client of the same type.

TransportClientNotFound

Unable to find a registered client for the given client context.

PeerAlreadyRegistered

Found an active client of the same type.

PeerNotFound

Unable to find the peer specified by the API.

FlowAlreadyEnabled

Flow already enabled

FlowNotFound

Flow not found

FlowConnectionUp

Trying to enable a flow that already has a valid connection

TransportClientConfigNotFound

Unable to find config for the client type.

TransportBusy(TransportPayload)

Failed to enqueue/submit a message/request. The error code contains the entry that could not be submitted.

ConnectionWriteFailed(String)

Write to connection failed due to OS error. The string has the human readable error string.

ConnectionReadFailed(String)

Read from connection failed due to OS error. The string has the human readable error string.

SerializationFailed

Failed to serialize

DeserializationFailed

Failed to deserialize

RegistryMissingConfig

Registry has missing entries for one of advert/request/artifact types

RegistryMultiIP

Registry has multiple IPs across advert/request/artifact types

RegistryInvalidPortNumber

Registry has invalid port number

MessageQueueRoutingFailed

Unable to route the message -> queue, based on the config.

TransportQueueFull

Transport queue is full.

TransportQueueStopped

The queue is being shut down.

ConnectionEventHandlerNotFound

Connection event handler not registered.

MessageSerializationFailed

Failed to serialize the message for send.

MessageConnectionRoutingFailed

Unable to route the message -> connection, based on the config.

ConnectionNotFound

Unable to find a connection to send the message.

SocketWriteFailed

Failed to write to socket.

ServerSocketCreateFailed

Server socket creation failed.

ServerSocketAddrReuseFailed

Failed to set the address reuse flag on the server socket.

ServerSocketPortReuseFailed

Failed to set the port reuse flag on the server socket.

ServerSocketBindFailed

Failed to bind to server port.

ServerSocketListenFailed

Failed to listen on the server socket.

ServerSocketConversionFailed

Failed to convert server listener.

ClientSocketCreateFailed

Client socket creation failed.

ClientSocketBindFailed

Client socket bind failed.

SocketNoDelayFailed

Failed to set the NO_DELAY option

ServerDown

Server not accepting connections on the specified port.

ConnectOsError

connect() failed with unhandled OS error.

RegistryDuplicateNodeId

Duplicate node Ids in node registry.

RegistryDuplicateNodeIP

Duplicate node IPs in node registry.

RegistryDuplicateEndpoint

Duplicate <node IP, port> in node registry.

RegistryInvalidNodeIP

Invalid IP address in node registry.

NodeIpResolutionFailed

NodeId -> IP resolution failed.

NodeServerEndpointResolutionFailed

NodeId -> server endpoint resolution failed.

WrapperCertParsingFailed

Failed to parse the PEM certificate

NodeIdMissing

NodeId missing from the certificate

NodeIdParsingFailed

Failed to parse the NodeId from the certificate

NodeIdMalformed

NodeId in the certificate was not in the expected format

DomainNameMissing

Domain name missing from the certificate

DomainNameTooMany

Too many domain names in the certificate

DomainNameMalformed

Domain name in the certificate was not in the expected format

PublicKeyParsingFailed

Failed to get the public key from the certificate

RegistryTlsConfigNotUniform

TLS is not uniformly configured across all the registry nodes

InvalidNodeIdInCertificate

The NodeId in the certificate is incorrect

PeerTlsInfoNotFound

Failed to find peer TLS info

PeerTlsInfoMismatch

Peer cert did not match the expected value in the registry

ConfigPrivateKeyParsingFailed

The private key file specified in the config could not be parsed

ConfigPrivateKeyFileReadFailed

The private key file specified in the config could not be read

ConfigCertParsingFailed

The private key file specified in the config could not be parsed

ConfigCertFileReadFailed

The private key file specified in the config could not be read

SetNodeKeyFailed

Failed to initialize the node key prior to the TLS handshake

SetNodeCertFailed

Failed to initialize the certificate prior to the TLS handshake

AddPeerCertFailed

Failed to add peer certificate prior to the TLS handshake

AcceptorInitFailed

Failed to initialize the acceptor

ConnectorInitFailed

Failed to initialize the connector

ConnectorConfigFailed

Failed to configure the client side TLS connector

SenderErrorIndicated

Received an error from sender

InvalidSockAddr

Failed to get socket address

NodeRecordDuplicateFlowTag

Duplicate flow tags in NodeRecord

NodeRecordMissingConnectionEndpoint

Missing connection endpoint in NodeRecord

TimeoutExpired

Timeout expired