Enum ic_types::user_error::ErrorCode

pub enum ErrorCode {
    SubnetOversubscribed,
    CanisterOutputQueueFull,
    IngressMessageTimeout,
    CanisterNotFound,
    CanisterMethodNotFound,
    CanisterAlreadyInstalled,
    CanisterWasmModuleNotFound,
    CanisterEmpty,
    InsufficientTransferFunds,
    InsufficientMemoryAllocation,
    InsufficientCyclesForCreateCanister,
    SubnetNotFound,
    CanisterOutOfCycles,
    CanisterTrapped,
    CanisterCalledTrap,
    CanisterContractViolation,
    CanisterInvalidWasm,
    CanisterDidNotReply,
    CanisterOutOfMemory,
    CanisterStopped,
    CanisterStopping,
    CanisterNotStopped,
    CanisterStoppingCancelled,
    CanisterInvalidController,
    CanisterFunctionNotFound,
    CanisterNonEmpty,
    CertifiedStateUnavailable,
    CanisterRejectedMessage,
    InterCanisterQueryLoopDetected,
    UnknownManagementMessage,
    InvalidManagementPayload,
    InsufficientCyclesInCall,
}

User-facing error codes.

The error codes are currently assigned using an HTTP-like convention: the most significant digit is the corresponding reject code and the rest is just a sequentially assigned two-digit number.

Variants

SubnetOversubscribed

CanisterOutputQueueFull

IngressMessageTimeout

CanisterNotFound

CanisterMethodNotFound

CanisterAlreadyInstalled

CanisterWasmModuleNotFound

CanisterEmpty

InsufficientTransferFunds

InsufficientMemoryAllocation

InsufficientCyclesForCreateCanister

SubnetNotFound

CanisterOutOfCycles

CanisterTrapped

CanisterCalledTrap

CanisterContractViolation

CanisterInvalidWasm

CanisterDidNotReply

CanisterOutOfMemory

CanisterStopped

CanisterStopping

CanisterNotStopped

CanisterStoppingCancelled

CanisterInvalidController

CanisterFunctionNotFound

CanisterNonEmpty

CertifiedStateUnavailable

CanisterRejectedMessage

InterCanisterQueryLoopDetected

UnknownManagementMessage

InvalidManagementPayload

InsufficientCyclesInCall