Module ic_types::xnet

Types used by the Xnet component.

Modules

proto
Protocol buffer equivalents to the various structs that make up a CertifiedStreamSlice, for backwards- and forwards-compatible XNet wire format.

testing

Structs

CertifiedStreamSlice
A slice of the stream of messages produced by the other subnet together with a cryptographic proof that the majority of the subnet agrees on it.

QueueId
A QueueId. Identifies a message queue by destination, source and session ID. Note that the tuple order is an important consideration as it supports appropriate clustering of outgoing message streams.

SessionTag

StreamHeader
StreamHeader contains a digest of information about the communication session between subnets: indices of messages that can be pulled and signals for received incoming messages.

StreamIndexTag

StreamIndexedQueue
A gap-free StreamIndex-ed queue for the messages and signals of a stream.

StreamSlice
A continuous slice of messages pulled from a remote subnet. The slice also includes the header with the communication session metadata.

Type Definitions

SessionId
Identifies a session between a given pair of sender,receiver canisters.

StreamIndex
Index into a subnet-to-subnet message stream; used in the context of a Stream to define message order.